Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Implemented functionality that allows users to see a list of available channels in teh chatroom so they can choose which one they wish to connect to. This required changes in websocket_handler to allow us to pass back the name of all the channels, a modification of the broadcast message in global_websocket_state to include the new information and changes to chat.html to be able to see these new changes. note these updates only happen a message is sent or recieved and does not automatically appear when a connection is setup. This could be addressed in a future CL.
Commit: 2006cb7d93e2c8dd7b804fe858278fed941acff7
Tag: autogenerated:gerrit:newPatchSet
Groups: f4e29b6208209811e478919cce39aff120de4a65